Refactor experiment config page to a participant-like view
This aligns the config page more closely with the participant's view of
the experiment and allows for more configuration room.
